+++ click_extended/decorators/misc/__init__.py
@@ -5,6 +5,7 @@ from click_extended.decorators.misc.confirm_if import confirm_if
 from click_extended.decorators.misc.default import default
 from click_extended.decorators.misc.deprecated import deprecated
 from click_extended.decorators.misc.experimental import experimental
+from click_extended.decorators.misc.now import now
 
 __all__ = [
     "choice",
@@ -12,4 +13,5 @@ __all__ = [
     "default",
     "deprecated",
     "experimental",
+    "now",
 ]